Updating subtitleService >> stremioService
